MovieQuizPresenter added. Moved questionsAmount & currentQuestionIndex to presenter file.
